feat: move some types to types package
