+++ packages/types/src/api-reference/index.ts
@@ -21,3 +21,5 @@ export type {
   SpecificationExtension,
   ApiReferencePlugin,
 } from './api-reference-plugin.ts'
+
+export type { ApiReferenceInstance, CreateApiReference, ScalarGlobal } from './html-api.ts'
